Refactor done, existing functionality maintained
