+++ src/components/Office/AccountingCodeSection/AccountingCodeSection.module.scss
@@ -22,7 +22,3 @@
     @include u-text('primary-dark');
   }
 }
-
-.addCodeBtn {
-  width: auto;
-}
