+++ src/components/Office/AccountingCodesModal/AccountingCodesModal.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const AccountingCodesModal = () => {
+  return <div>hey</div>;
+};
+
+export default AccountingCodesModal;
