+++ src/components/Office/ShipmentAccountingCodes/ShipmentAccountingCodes.module.scss
@@ -1,0 +1,3 @@
+.addCodeBtn {
+  width: auto;
+}
