+++ src/components/Office/ShipmentAccountingCodes/ShipmentAccountingCodes.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Grid, GridContainer } from '@trussworks/react-uswds';
 import { Formik } from 'formik';
 
-import AccountingCodes from './AccountingCodes';
+import ShipmentAccountingCodes from './ShipmentAccountingCodes';
 
 import styles from 'pages/Office/ServicesCounselingMoveInfo/ServicesCounselingTab.module.scss';
 import shipmentFormStyles from 'components/Office/ShipmentForm/ShipmentForm.module.scss';
@@ -10,8 +10,8 @@ import { Form } from 'components/form/Form';
 import formStyles from 'styles/form.module.scss';
 
 export default {
-  title: 'Office Components / Forms / ShipmentForm / AccountingCodes',
-  component: AccountingCodes,
+  title: 'Office Components / Forms / ShipmentForm / ShipmentAccountingCodes',
+  component: ShipmentAccountingCodes,
   decorators: [
     (Story) => (
       <GridContainer className={styles.gridContainer}>
@@ -37,25 +37,25 @@ export default {
 
 export const AsRequired = () => (
   <div className="officeApp">
-    <AccountingCodes optional={false} />{' '}
+    <ShipmentAccountingCodes optional={false} />{' '}
   </div>
 );
 
 export const WithNoTACsOrSACs = () => (
   <div className="officeApp">
-    <AccountingCodes />
+    <ShipmentAccountingCodes />
   </div>
 );
 WithNoTACsOrSACs.storyName = 'With No TACs or SACs';
 
 export const WithSingleCode = () => (
   <div className="officeApp">
-    <AccountingCodes TACs={{ HHG: '1234', NTS: undefined }} />
+    <ShipmentAccountingCodes TACs={{ HHG: '1234', NTS: undefined }} />
   </div>
 );
 
 export const WithMultipleCodes = () => (
   <div className="officeApp">
-    <AccountingCodes TACs={{ HHG: '1234', NTS: '5678' }} SACs={{ HHG: '98765', NTS: '000012345' }} />
+    <ShipmentAccountingCodes TACs={{ HHG: '1234', NTS: '5678' }} SACs={{ HHG: '98765', NTS: '000012345' }} />
   </div>
 );
